Change AbstractBlob.equal() to base off equality of blob id.
Since the identity of a blob is a hash of the content of the blob
(at least this is the implementation currently), we can gain
performance benefit in AbstractBlob.equal() by relying on the
equality state of the blob id to be the same as the equality
state of the blob itself, if both blob ids are available
(meaning, not null).

With this implementation, blob stores will only check the
equality state of the blobs themselves if one or both blob
ids are null.  This matters especially for cloud blob stores.

However, if the implementation of creating blob ids were
to ever change from being a content hash, this optimization
will no longer be valid.

Issue: [#OAK-5253]