Always return a writable delegate when one is asked for.
If a data identifier is provided when requesting a writable
delegate, we should prefer the delegate that contains a match
for the identifier, if there is one.  But if no delegate contains
a match, we should return the first writable delegate if there
is one.  This commit fixes this bug; before, if no match could
be found no delegate would be returned even if there was a
writable delegate that could be used.

Issue: [#GRANITE-18236]