Separate data store initiation and registration for factory class.
If a data store is being created via a service class, the service
class needs to initialize and register the data store.  If a
data store is being created via a factory class, for example as
a delegate data store for CompositeDataStore, the data store
(delegate) needs to be initialized, but not registered.  Instead
the delegate will be registered via the CompositeDataStoreService
as a DataStoreProvider and the CompositeDataStore itself will
be registered as the data store.

This change separates the init and register logic in
AbstractDataStoreService, keeping both static so we can reuse code
from the factory class.  The factory only calls init, not
register.  This change also slightly modifies
CompositeDataStoreService so it leverages the change and works
with deferred registrations, especially in the cases where
delegates get registered before the CompositeDataStore.