Disable filename* portion of requested Content-Disposition header.
Azure's SDK appears to have a bug in how it handles the encoded
portion of the requested Content-Disposition header; see
https://github.com/Azure/azure-sdk-for-java/issues/2900.  Encoding
the value for the filename* portion of the header causes requests
to the resulting URI to fail with 403 errors; failing to encode
the filename* portion results in an improperly-formatted
Content-Disposition header per RFC-8187.

This change disables the filename* portion until the SDK is fixed.
This will have the effect of making filenames outside the
ISO-8859-1 character set unsupported for now, but filenames with
spaces will work.