Add support for Azure CDNs for blob downloads in AzureDataStore.
In Azure it is possible to configure a CDN for a storage account.
If configured, blobs can be downloaded from the blob storage
container via a CDN URI for faster download speeds.  This change
adds a configuration option "cdnDomain" to the AzureDataStore
configuration.   If this optional configuration value is set,
the value will be used as the domain instead of the Azure default
"<storage-account-name>.blob.core.windows.net" when creating
direct download URIs.