Update direct binary access doc relating to initiateHttpUpload().
The user documentation for direct binary access claimed that
the result of calling initiateHttpUpload() could be a null return
value, or a BinaryUpload with no URIs inside, either situation
indicating that the direct binary upload feature is not enabled.
This disagrees somewhat with the Javadoc which only cites a null
return value as the indicator that the feature is not enabled.
This commit modifies that line in the documentation to match
the Javadoc, only specifying that a null return value indicates
that the feature is not enabled.