Fix signed URI generation and record deletion bugs.
Signed URI:  Ran into intermittent test failures due to having
improperly encoded signatures for certain URIs (usually worked but
not always).  Constructing a URI by parts doesn't work well;
this update changes it to construct a URI by parsing the full URI
string, which doesn't modify the query parameter encoding.
Record deletion:  If the record doesn't exist (which is ok), the
Azure API throws a StorageException.  This change catches the
exception and checks for a 404 status; if the status is 404, we
simply log a message and don't throw a DataStoreException.