Update docs to show DBA disabled if using SAS authentication.
For AzureDataStore, it is not possible to generate signed URIs if
the data store is configured to authenticate using a shared access
signature.  This commit updates the documentation to note this
limitation and to indicate that the direct binary access feature
is disabled if the AzureDataStore authenticates using a shared
access signature.  To use direct binary access, AzureDataStore
authentication must be configured to use an access key.